requirements.txt
PyQt5>=5.15.0
librosa>=0.10.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
soundfile>=0.12.0

